pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "chromadb",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-chroma",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq",
    "numpy<2",
    "pymupdf>=1.27.2",
    "pypdf",
    "python-dotenv",
    "pyyaml>=6.0.3",
    "streamlit>=1.55.0",
]
